renames the cmdb package to infocmdb to be conform to the naming
