+++ infocmdb/v1/infocmdb/cmdb_test.go
@@ -1,4 +1,4 @@
-package cmdb
+package infocmdb
 
 import (
 	"bytes"
