+++ infocmdb/v1/infocmdb/webservice.go
@@ -1,4 +1,4 @@
-package cmdb
+package infocmdb
 
 import (
 	"bytes"
