+++ infocmdb/v1/infocmdb/workflow.go
@@ -1,4 +1,4 @@
-package cmdb
+package infocmdb
 
 import (
 	"encoding/json"
