+++ infocmdb/v2/infocmdb/ci.go
@@ -1,8 +1,8 @@
-package cmdb
+package infocmdb
 
 import (
 	"fmt"
-	"github.com/infonova/infocmdb-sdk-go/infocmdb/v2/cmdb/client"
+	"github.com/infonova/infocmdb-sdk-go/infocmdb/v2/infocmdb/client"
 	"gopkg.in/resty.v1"
 	"strconv"
 )
