+++ infocmdb/v2/infocmdb/ci_test.go
@@ -1,4 +1,4 @@
-package cmdb
+package infocmdb
 
 import (
 	"testing"
