+++ infocmdb/v2/infocmdb/cmdb.go
@@ -1,37 +1,38 @@
-package cmdb
+package infocmdb
 
 import (
-	"errors"
 	"io/ioutil"
+	"net/http"
+	"net/url"
 	"os"
 	"path/filepath"
 	"time"
 
+	"github.com/infonova/infocmdb-sdk-go/infocmdb/v2/infocmdb/client"
 	"github.com/patrickmn/go-cache"
 	log "github.com/sirupsen/logrus"
 	"gopkg.in/yaml.v2"
 )
 
-var (
-	ErrFailedToCreateInfoCMDB  = errors.New("failed to create infocmdb object")
-	ErrNoCredentials           = errors.New("must provide credentials")
-	ErrNotImplemented          = errors.New("not implemented")
-	ErrNoResult                = errors.New("query returned no result")
-	ErrTooManyResults          = errors.New("query returned to many results, expected one")
-	ErrWebserviceResponseNotOk = errors.New("webservice response was not ok")
-)
-
 type Config struct {
-	ApiUrl       string `yaml:"apiUrl"`
-	ApiUser      string `yaml:"apiUser"`
-	ApiPassword  string `yaml:"apiPassword"`
-	ApiKey       string
-	CmdbBasePath string `yaml:"CmdbBasePath"`
+	Url      string `yaml:"apiUrl"`
+	Username string `yaml:"apiUser"`
+	Password string `yaml:"apiPassword"`
+	ApiKey   string
+	BasePath string `yaml:"BasePath"`
 }
 
 type Cmdb struct {
 	Config Config
 	Cache  *cache.Cache
+	Client *client.Client
+	Logger *log.Logger
+	Error  error
+}
+
+type ErrorReturn struct {
+	Message string `json:"message"`
+	Success bool   `json:"success"`
 }
 
 type CiRelationDirection string
@@ -53,6 +54,15 @@ const (
 	ATTRIBUTE_VALUE_TYPE_CI                         = "value_ci"
 )
 
+type UpdateMode string
+
+const (
+	UPDATE_MODE_INSERT UpdateMode = "insert"
+	UPDATE_MODE_UPDATE            = "update"
+	UPDATE_MODE_DELETE            = "delete"
+	UPDATE_MODE_SET               = "set"
+)
+
 func init() {
 	log.SetLevel(log.InfoLevel)
 	if os.Getenv("WORKFLOW_DEBUGGING") == "true" {
@@ -62,12 +72,12 @@ func init() {
 
 func (i *Cmdb) LoadConfigFile(configFile string) (err error) {
 	_, err = os.Stat(configFile)
-	if err == nil {
-		log.Debugf("ConfigFile found with given string: %s", configFile)
-	} else {
+	if os.IsNotExist(err) {
 		WorkflowConfigPath := os.Getenv("WORKFLOW_CONFIG_PATH")
 		log.Debugf("WORKFLOW_CONFIG_PATH: %s", WorkflowConfigPath)
 		configFile = filepath.Join(WorkflowConfigPath, configFile)
+	} else if err != nil {
+		return
 	}
 
 	log.Debugf("ConfigFile: %s", configFile)
@@ -82,34 +92,50 @@ func (i *Cmdb) LoadConfigFile(configFile string) (err error) {
 		return
 	}
 
-	return i.LoadConfig(yamlFile)
+	err = i.LoadConfig(yamlFile)
+	return
 }
 
 func (i *Cmdb) LoadConfig(config []byte) (err error) {
-	return yaml.Unmarshal(config, &i.Config)
-}
+	if err = yaml.Unmarshal(config, &i.Config); err != nil {
+		return
+	}
 
-func New(config string) (i *Cmdb, err error) {
-	i = new(Cmdb)
-	err = i.LoadConfigFile(config)
+	err = i.applyUrlFromRedirect()
 	if err != nil {
-		log.Error(err)
-		return nil, err
+		return
 	}
 
-	i.Cache = cache.New(5*time.Minute, 10*time.Minute)
-
-	return i, nil
+	i.Client = client.New(i.Config.Url)
+	return
 }
 
-func (i *Cmdb) Login() error {
-	if i.Config.ApiKey != "" {
-		log.Debug("already logged in")
-		return nil
+func (i *Cmdb) applyUrlFromRedirect() (err error) {
+	c := &http.Client{
+		CheckRedirect: func(req *http.Request, via []*http.Request) error {
+			return http.ErrUseLastResponse
+		},
+	}
+	resp, err := c.Get(i.Config.Url)
+	if err != nil {
+		return
 	}
 
-	if i.Config.ApiUser == "" {
-		return ErrNoCredentials
+	for name, header := range resp.Header {
+		if name == "Location" && len(header) > 0 {
+			baseUrl, _ := url.Parse(header[0])
+			i.Config.Url = baseUrl.Scheme + "://" + baseUrl.Host + "/"
+		}
 	}
-	return i.LoginWithUserPass(i.Config.ApiUrl, i.Config.ApiUser, i.Config.ApiPassword)
+
+	return
+}
+
+// New returns a new Cmdb Client to access the V2 Api
+func New() (i *Cmdb) {
+	return &Cmdb{
+		Cache:  cache.New(5*time.Minute, 10*time.Minute),
+		Client: &client.Client{},
+		Logger: log.New(),
+	}
 }
