+++ infocmdb/v2/infocmdb/query_test.go
@@ -1,9 +1,9 @@
-package cmdb
+package infocmdb
 
 import (
 	"testing"
 
-	"github.com/infonova/infocmdb-sdk-go/infocmdb/v2/cmdb/client"
+	"github.com/infonova/infocmdb-sdk-go/infocmdb/v2/infocmdb/client"
 	utilTesting "github.com/infonova/infocmdb-sdk-go/util/testing"
 	"github.com/patrickmn/go-cache"
 )
