+++ infocmdb/webservice_default_test.go
@@ -5,8 +5,8 @@ import (
 	"reflect"
 	"testing"
 
-	v1 "github.com/infonova/infocmdb-sdk-go/infocmdb/v1/cmdb"
-	v2 "github.com/infonova/infocmdb-sdk-go/infocmdb/v2/cmdb"
+	v1 "github.com/infonova/infocmdb-sdk-go/infocmdb/v1/infocmdb"
+	v2 "github.com/infonova/infocmdb-sdk-go/infocmdb/v2/infocmdb"
 	"github.com/joho/godotenv"
 	log "github.com/sirupsen/logrus"
 )
